backend/requirements.txt
# Core dependencies
uvicorn==0.34.2
fastapi==0.109.2
pydantic==2.11.5
python-dotenv==1.1.0

# Vector database and embeddings
qdrant-client==1.14.2
numpy<2.0.0
sentence-transformers==4.1.0
transformers>=4.41.0,<5.0.0
langchain-core>=0.3.13,<0.4.0
langchain-qdrant>=0.1.1
langchain-huggingface>=0.1.1
openai>=1.0.0
httpx>=0.24.1

# Document processing
PyPDF2==3.0.1  
python-docx==1.1.2
python-pptx==1.0.2
langchain-text-splitters>=0.3.8

# Storage
boto3==1.34.34
botocore==1.34.34
requests==2.32.3

# Additional dependencies
torch>=2.1.0
tiktoken==0.6.0
accelerate==0.27.2 